requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
email-validator
httpx